refactor: util 클래스의 인스턴스 생성을 막기 위해 private 생성자 선언
