+++ src/main/java/lotto/util/LottoNumberGenerator.java
@@ -1,4 +1,6 @@
-package lotto.model;
+package lotto.util;
+
+import lotto.model.LottoNumber;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -7,6 +9,9 @@ import static lotto.model.Lotto.CANDIDATE_LOTTO_NUMBER_SIZE;
 import static lotto.model.Lotto.CORRECT_LOTTO_NUMBERS_SIZE;
 
 public class LottoNumberGenerator {
+
+    private LottoNumberGenerator() {
+    }
 
     public static Set<LottoNumber> makeLottoNumber() {
         Set<LottoNumber> lotto = new HashSet<>();
