+++ src/test/java/lotto/model/LottoTest.java
@@ -1,5 +1,6 @@
 package lotto.model;
 
+import lotto.util.LottoNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
